CS 554 Big Data Project/ecommerce-bigdata-pipeline/dashboard/requirements.txt
dash==2.14.2
plotly==5.18.0
kafka-python==2.0.2
happybase==1.2.0
